+++ ml/backend/ggml/ggml/src/ggml-cuda/getrows.cu
@@ -299,6 +299,14 @@ static void ggml_cuda_get_rows_switch_src0_type(
             get_rows_cuda_q<QK8_0, QR8_0, dequantize_q8_0>(src0_d, src1_d, dst_d,
                 ne00, nb01, nb02, nb03, ne10, ne11, ne12, nb10, nb11, nb12, nb1, nb2, nb3, stream);
             break;
+        case GGML_TYPE_FP8_E4M3:
+            get_rows_cuda_q<QK_FP8_E4M3, QR_FP8_E4M3, dequantize_fp8_e4m3>(src0_d, src1_d, dst_d,
+                ne00, nb01, nb02, nb03, ne10, ne11, ne12, nb10, nb11, nb12, nb1, nb2, nb3, stream);
+            break;
+        case GGML_TYPE_FP8_E5M2:
+            get_rows_cuda_q<QK_FP8_E5M2, QR_FP8_E5M2, dequantize_fp8_e5m2>(src0_d, src1_d, dst_d,
+                ne00, nb01, nb02, nb03, ne10, ne11, ne12, nb10, nb11, nb12, nb1, nb2, nb3, stream);
+            break;
         case GGML_TYPE_Q6_K:
             get_rows_cuda_q6_K(src0_d, src1_d, dst_d,
                 ne00, nb01, nb02, nb03, ne10, ne11, ne12, nb10, nb11, nb12, nb1, nb2, nb3, stream);